Decouple culling tests
Extracted some culling related functionality to a new module. Like
this, tests dont import from other test suites.
