+++ tests/test_culling.py
@@ -9,28 +9,17 @@ from unittest.mock import patch
 import pytest
 
 from app import db
-from app.models import Host
 from app.utils import HostWrapper
 from host_reaper import run as host_reaper_run
 from tests.test_api_delete import DeleteHostsBaseTestCase
 from tests.test_api_utils import ACCOUNT
-from tests.test_api_utils import ApiBaseTestCase
 from tests.test_api_utils import DbApiTestCase
 from tests.test_api_utils import generate_uuid
 from tests.test_api_utils import HOST_URL
 from tests.test_api_utils import now
 from tests.test_api_utils import test_data
+from tests.test_culling_utils import HostStalenessBaseTestCase
 from tests.test_utils import MockEventProducer
-
-
-class CullingBaseTestCase(ApiBaseTestCase):
-    def _nullify_culling_fields(self, host_id):
-        with self.app.app_context():
-            host = db.session.query(Host).get(host_id)
-            host.stale_timestamp = None
-            host.reporter = None
-            db.session.add(host)
-            db.session.commit()
 
 
 class QueryStaleTimestampTestCase(DbApiTestCase):
@@ -75,22 +64,7 @@ class QueryStaleTimestampTestCase(DbApiTestCase):
         _assert_values(get_by_id_response["results"][0])
 
 
-class QueryStalenessBaseTestCase(DbApiTestCase):
-    def _create_host(self, stale_timestamp):
-        data = {
-            "account": ACCOUNT,
-            "insights_id": str(generate_uuid()),
-            "facts": [{"facts": {"fact1": "value1"}, "namespace": "ns1"}],
-        }
-        if stale_timestamp:
-            data["reporter"] = "some reporter"
-            data["stale_timestamp"] = stale_timestamp.isoformat()
-
-        host = HostWrapper(data)
-        response = self.post(HOST_URL, [host.data()], 207)
-        self._verify_host_status(response, 0, 201)
-        return self._pluck_host_from_response(response, 0)
-
+class QueryHostsStalenessBaseTestCase(HostStalenessBaseTestCase):
     def _get_hosts_by_id_url(self, host_id_list, endpoint="", query=""):
         host_id_query = ",".join(host_id_list)
         return f"{HOST_URL}/{host_id_query}{endpoint}{query}"
@@ -101,18 +75,7 @@ class QueryStalenessBaseTestCase(DbApiTestCase):
         return response["results"]
 
 
-class QueryStalenessGetHostsBaseTestCase(QueryStalenessBaseTestCase, CullingBaseTestCase):
-    def setUp(self):
-        super().setUp()
-        current_timestamp = now()
-        self.fresh_host = self._create_host(current_timestamp + timedelta(hours=1))
-        self.stale_host = self._create_host(current_timestamp - timedelta(hours=1))
-        self.stale_warning_host = self._create_host(current_timestamp - timedelta(weeks=1, hours=1))
-        self.culled_host = self._create_host(current_timestamp - timedelta(weeks=2, hours=1))
-
-        self.unknown_host = self._create_host(current_timestamp)
-        self._nullify_culling_fields(self.unknown_host["id"])
-
+class QueryStalenessGetHostsBaseTestCase(QueryHostsStalenessBaseTestCase, HostStalenessBaseTestCase):
     def _created_hosts(self):
         return (self.fresh_host["id"], self.stale_host["id"], self.stale_warning_host["id"], self.culled_host["id"])
 
@@ -236,7 +199,7 @@ class QueryStalenessGetHostsIgnoresStalenessParameterTestCase(QueryStalenessGetH
         self._check_query_fails("/system_profile", "?staleness=fresh")
 
 
-class QueryStalenessConfigTimestampsTestCase(QueryStalenessBaseTestCase):
+class QueryStalenessConfigTimestampsTestCase(QueryHostsStalenessBaseTestCase):
     def _create_and_get_host(self, stale_timestamp):
         host_to_create = self._create_host(stale_timestamp)
         retrieved_host = self._get_hosts_by_id((host_to_create["id"],))[0]
@@ -268,7 +231,7 @@ class QueryStalenessConfigTimestampsTestCase(QueryStalenessBaseTestCase):
                 self.assertEqual(culled_timestamp.isoformat(), host["culled_timestamp"])
 
 
-class HostReaperTestCase(DeleteHostsBaseTestCase, CullingBaseTestCase):
+class HostReaperTestCase(DeleteHostsBaseTestCase):
     def setUp(self):
         super().setUp()
         self.now_timestamp = datetime.now(timezone.utc)
@@ -343,8 +306,6 @@ class HostReaperTestCase(DeleteHostsBaseTestCase, CullingBaseTestCase):
             added_hosts = self._add_hosts(({},))
             added_host_id = added_hosts[0].id
             self._check_hosts_are_present((added_host_id,))
-
-            self._nullify_culling_fields(added_host_id)
 
             self._run_host_reaper()
             self._check_hosts_are_present((added_host_id,))
